Make the scrolling more imperative
